requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
datasets==3.1.0
dill==0.3.8
distro==1.9.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
numpy==2.1.1
openai==1.50.2
packaging==24.1
pandas==2.2.3
patsy==1.0.1
propcache==0.2.0
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
statsmodels==0.14.4
sympy==1.13.3
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.18.0
